requirements.txt
PyMuPDF
chromadb
langchain
langchain-community
langchain-google-genai
python-dotenv